Реализация Service слоя для User
