+++ Chat/Service Layer/Assembly/ServicesAssembly.swift
@@ -1,0 +1,34 @@
+//
+//  ServicesAssembly.swift
+//  Chat
+//
+//  Created by Алексей Никитин on 09.11.2020.
+//  Copyright © 2020 Алексей Никитин. All rights reserved.
+//
+
+import Foundation
+
+enum AllowedUserStorages {
+    case gcd, operations
+}
+
+protocol ServicesAssemblyProtocol {
+    func getUserService(for userStorage: AllowedUserStorages) -> UserServiceProtocol
+}
+
+class ServicesAssembly: ServicesAssemblyProtocol {
+    private let coreAssembly: CoreAssemblyProtocol
+    
+    init(coreAssembly: CoreAssemblyProtocol) {
+        self.coreAssembly = coreAssembly
+    }
+    
+    func getUserService(for userStorage: AllowedUserStorages) -> UserServiceProtocol {
+        switch userStorage {
+        case .gcd:
+            return UserService(storage: coreAssembly.userCGDStorage)
+        case .operations:
+            return UserService(storage: coreAssembly.userOperationsStorage)
+        }
+    }
+}
